+++ src/renderer/delta-remote.ts
@@ -307,6 +307,7 @@ class DeltaRemote {
     fromServer: boolean,
     seconds: number
   ): Promise<number>
+  call(fnName: 'settings.setConfigFromQr', qrCodeContent: string): Promise<void>
   // stickers -----------------------------------------------------------
   call(
     fnName: 'stickers.getStickers'